+++ frontend/src/employee-frontend/App.tsx
@@ -21,7 +21,7 @@ import EmployeeRoute from './components/EmployeeRoute'
 import FinancePage from './components/FinancePage'
 import { Footer } from './components/Footer'
 import Header from './components/Header'
-import IncomeStatementPage from './components/IncomeStatementPage'
+import IncomeStatementPage from './components/income-statements/IncomeStatementPage'
 import LoginPage from './components/LoginPage'
 import MobilePairingModal from './components/MobilePairingModal'
 import PersonalMobileDevicesPage from './components/PersonalMobileDevicesPage'
